@@ -1,12 +1,12 @@
 <Project>
     <PropertyGroup>
-        <TargetFrameworks>net6.0;net7.0;net8.0</TargetFrameworks> <!-- LTS | STS | LTS -->
-        <TestDevelopmentVersion>17.9.0</TestDevelopmentVersion> <!-- Microsoft.NET.Test.Sdk -->
-        <TestFrameworkVersion>3.3.1</TestFrameworkVersion> <!-- MSTest.Test -->
+        <TargetFrameworks>net8.0;net9.0</TargetFrameworks>
 
-        <CommandLibraryVersion>2.9.1</CommandLibraryVersion> <!-- CommandLineParser -->
-        <DatabaseLibraryVersion>5.2.0</DatabaseLibraryVersion> <!-- Microsoft.Data.SqlClient -->
-        <TimeLibraryVersion>3.1.11</TimeLibraryVersion> <!-- NodaTime -->
+        <TestDevelopmentVersion>17.11.1</TestDevelopmentVersion> <!-- Microsoft.NET.Test.Sdk -->
+        <TestFrameworkVersion>3.6.3</TestFrameworkVersion> <!-- MSTest.Test -->
+
+        <DatabaseLibraryVersion>5.2.2</DatabaseLibraryVersion> <!-- Microsoft.Data.SqlClient -->
+        <TimeLibraryVersion>3.2.0</TimeLibraryVersion> <!-- NodaTime -->
         <TransitLibraryVersion>3.0.0-pre006</TransitLibraryVersion> <!-- GTFS -->
     </PropertyGroup>
 </Project>
