@@ -7,17 +7,17 @@ namespace NextDepartures.Standard.Extensions
 {
     public static class StringExtensions
     {
-        private static readonly string[] separator = new string[] { ":" };
+        private static readonly string[] Separator = [":"];
 
         public static TimeOfDay? ToTimeOfDay(this string baseString)
         {
-            int[] splittedString = baseString.Split(separator, StringSplitOptions.None).Select(s => int.Parse(s)).ToArray();
+            var splitTime = baseString.Split(Separator, StringSplitOptions.None).Select(int.Parse).ToArray();
 
             return new TimeOfDay()
             {
-                Hours = splittedString[0],
-                Minutes = splittedString[1],
-                Seconds = splittedString[2]
+                Hours = splitTime[0],
+                Minutes = splitTime[1],
+                Seconds = splitTime[2]
             };
         }
 
@@ -28,7 +28,7 @@ namespace NextDepartures.Standard.Extensions
 
         public static string WithPrefix(this string baseString, string prefix)
         {
-            return string.Format("{0}{1}", prefix, baseString);
+            return $"{prefix}{baseString}";
         }
     }
 }
