@@ -8,10 +8,7 @@ namespace NextDepartures.Standard.Extensions
     {
         public static DateTime ToZonedDateTime(this DateTime dateTime, string timezone)
         {
-            Instant instant = dateTime.ToUniversalTime().ToInstant();
-            ZonedDateTime zoned = instant.InZone(DateTimeZoneProviders.Tzdb[timezone]);
-
-            return zoned.ToDateTimeUnspecified();
+            return dateTime.ToUniversalTime().ToInstant().InZone(DateTimeZoneProviders.Tzdb[timezone]).ToDateTimeUnspecified();
         }
     }
 }
