@@ -6,30 +6,15 @@ namespace NextDepartures.Standard.Extensions
     {
         public static int? ToInt32(this LocationType locationType)
         {
-            if (locationType == LocationType.Stop)
+            return locationType switch
             {
-                return 0;
-            }
-            else if (locationType == LocationType.Station)
-            {
-                return 1;
-            }
-            else if (locationType == LocationType.EntranceExit)
-            {
-                return 2;
-            }
-            else if (locationType == LocationType.GenericNode)
-            {
-                return 3;
-            }
-            else if (locationType == LocationType.BoardingArea)
-            {
-                return 4;
-            }
-            else
-            {
-                return null;
-            }
+                LocationType.Stop => 0,
+                LocationType.Station => 1,
+                LocationType.EntranceExit => 2,
+                LocationType.GenericNode => 3,
+                LocationType.BoardingArea => 4,
+                _ => null
+            };
         }
     }
 }
