@@ -8,16 +8,16 @@ namespace NextDepartures.Standard
     public partial class Feed
     {
         /// <summary>
-        /// Gets the agencies by the given email.
+        /// Gets the agencies by the given fare url.
         /// </summary>
-        /// <param name="email">The email. Default is all but can be overridden.</param>
+        /// <param name="fareUrl">The fare url. Default is all but can be overridden.</param>
         /// <param name="count">The maximum number of results to return. Default is all (0) but can be overridden.</param>
         /// <returns>A list of agencies.</returns>
-        public async Task<List<Agency>> GetAgenciesByEmailAsync(string email = "", int count = 0)
+        public async Task<List<Agency>> GetAgenciesByFareUrlAsync(string fareUrl = "", int count = 0)
         {
             try
             {
-                var agenciesFromStorage = await _dataStorage.GetAgenciesByEmailAsync(email);
+                var agenciesFromStorage = await _dataStorage.GetAgenciesByFareUrlAsync(fareUrl);
                 return count > 0 ? agenciesFromStorage.Take(count).ToList() : agenciesFromStorage;
             }
             catch
