@@ -1,5 +1,4 @@
-﻿using GTFS.Entities;
-using NextDepartures.Standard.Extensions;
+﻿using NextDepartures.Standard.Extensions;
 using NextDepartures.Standard.Models;
 using System;
 using System.Collections.Generic;
@@ -17,9 +16,39 @@ namespace NextDepartures.Standard
         /// <param name="hours">The maximum number of hours to search over. Default is all (0) but can be overridden.</param>
         /// <param name="count">The maximum number of results to return. Default is all (0) but can be overridden.</param>
         /// <returns>A list of services.</returns>
-        public Task<List<Service>> GetServicesByTripAsync(string id, int hours = 0, int count = 0)
+        public async Task<List<Service>> GetServicesByTripAsync(string id, int hours = 0, int count = 0)
         {
-            return GetServicesByTripAsync(id, DateTime.Now, TimeSpan.Zero, hours, count);
+            try
+            {
+                var agencies = await _dataStorage.GetAgenciesAsync();
+                var calendarDates = await _dataStorage.GetCalendarDatesAsync();
+                var departuresFromStorage = await _dataStorage.GetDeparturesForTripAsync(id);
+                var stops = await _dataStorage.GetStopsAsync();
+
+                List<Departure> departuresForTrip = [];
+
+                departuresForTrip.AddRange(new List<Departure>()
+                    .AddMultiple(GetDeparturesOnDay(agencies, calendarDates, stops, departuresFromStorage, DateTime.Now, DayOffsetType.Yesterday, TimeSpan.Zero, hours, id))
+                    .AddMultiple(GetDeparturesOnDay(agencies, calendarDates, stops, departuresFromStorage, DateTime.Now, DayOffsetType.Today, TimeSpan.Zero, hours, id))
+                    .AddMultiple(GetDeparturesOnDay(agencies, calendarDates, stops, departuresFromStorage, DateTime.Now, DayOffsetType.Tomorrow, TimeSpan.Zero, hours, id))
+                );
+
+                if (count > 0)
+                {
+                    return departuresForTrip
+                        .Take(count)
+                        .Select(d => CreateService(agencies, stops, d, "trip"))
+                        .ToList();
+                }
+
+                return departuresForTrip
+                    .Select(d => CreateService(agencies, stops, d, "trip"))
+                    .ToList();
+            }
+            catch
+            {
+                return null;
+            }
         }
 
         /// <summary>
@@ -35,12 +64,12 @@ namespace NextDepartures.Standard
         {
             try
             {
-                List<Agency> agencies = await _dataStorage.GetAgenciesAsync();
-                List<CalendarDate> calendarDates = await _dataStorage.GetCalendarDatesAsync();
-                List<Departure> departuresFromStorage = await _dataStorage.GetDeparturesForTripAsync(id);
-                List<Stop> stops = await _dataStorage.GetStopsAsync();
+                var agencies = await _dataStorage.GetAgenciesAsync();
+                var calendarDates = await _dataStorage.GetCalendarDatesAsync();
+                var departuresFromStorage = await _dataStorage.GetDeparturesForTripAsync(id);
+                var stops = await _dataStorage.GetStopsAsync();
 
-                List<Departure> departuresForTrip = new();
+                List<Departure> departuresForTrip = [];
 
                 departuresForTrip.AddRange(new List<Departure>()
                     .AddMultiple(GetDeparturesOnDay(agencies, calendarDates, stops, departuresFromStorage, now, DayOffsetType.Yesterday, offset, hours, id))
@@ -55,12 +84,10 @@ namespace NextDepartures.Standard
                         .Select(d => CreateService(agencies, stops, d, "trip"))
                         .ToList();
                 }
-                else
-                {
-                    return departuresForTrip
-                        .Select(d => CreateService(agencies, stops, d, "trip"))
-                        .ToList();
-                }
+
+                return departuresForTrip
+                    .Select(d => CreateService(agencies, stops, d, "trip"))
+                    .ToList();
             }
             catch
             {
